+++ historyAPI/frontend/index.html
@@ -16,6 +16,7 @@
     <div id="app"></div>
 </div>
 
-<script type="module" src="./index.js" defer></script>
+<script src="/static/index.js" type="module" defer></script>
 </body>
+
 </html>
