+++ historyAPI/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "historyapi",
+  "version": "1.0.0",
+  "description": "",
+  "main": "frontend/static/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.18.2"
+  }
+}
